refactor the project files structure
